Refactor smoke test to use yaml file and refactor package names
