+++ build.gradle
@@ -76,6 +76,10 @@ tasks.withType(Test) {
   }
 }
 
+tasks.withType(Copy).all {
+  duplicatesStrategy 'exclude'
+}
+
 test {
   failFast = true
 }
@@ -178,6 +182,25 @@ ext {
     springCloudVersion = "2021.0.2"
 }
 
+def versions = [
+  reformLogging: '5.1.9',
+  junit: '5.8.2',
+  junitPlatform   : '1.8.2',
+
+]
+ext['spring-security.version'] = '5.7.1'
+
+ext.libraries = [
+  "rest-assured": { version = '4.3.1' },
+  junit5: [
+    "org.junit.jupiter:junit-jupiter-api:${versions.junit}",
+    "org.junit.jupiter:junit-jupiter-engine:${versions.junit}",
+    "org.junit.jupiter:junit-jupiter-params:${versions.junit}",
+    "org.junit.platform:junit-platform-commons:${versions.junitPlatform}",
+    "org.junit.platform:junit-platform-engine:${versions.junitPlatform}"
+  ]
+]
+
 dependencyManagement {
   imports {
     mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
@@ -209,9 +232,18 @@ dependencies {
   implementation group: 'org.apache.logging.log4j', name: 'log4j-api', version: log4JVersion
   implementation group: 'org.apache.logging.log4j', name: 'log4j-to-slf4j', version: log4JVersion
   implementation group: 'io.rest-assured', name: 'rest-assured'
+  implementation 'junit:junit:4.13.1'
 
+/*
   testImplementation(platform('org.junit:junit-bom:5.8.2'))
   testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
+  testImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', {
+    exclude group: 'junit', module: 'junit'
+    exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
+  }
+*/
+
+  testImplementation libraries.junit5
   testImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', {
     exclude group: 'junit', module: 'junit'
     exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
@@ -219,6 +251,7 @@ dependencies {
 }
 
 mainClassName = 'uk.gov.hmcts.legacy.web.Application'
+
 
 bootJar {
   getArchiveFileName().set(provider {
