+++ src/smokeTest/resources/application-smoke.yaml
@@ -1,0 +1,1 @@
+test-url: http://localhost:8080
